move InfoMessage component from atoms to molecules
