+++ app/src/molecules/InfoMessage/__tests__/InfoMessage.test.tsx
