+++ app/src/molecules/InfoMessage/index.tsx
