+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunSetup.tsx
@@ -5,7 +5,7 @@ import { Flex, DIRECTION_COLUMN, SPACING } from '@opentrons/components'
 import { protocolHasModules } from '@opentrons/shared-data'
 
 import { Line } from '../../../atoms/structure'
-import { InfoMessage } from '../../../atoms/InfoMessage'
+import { InfoMessage } from '../../../molecules/InfoMessage'
 import {
   useDeckCalibrationData,
   useProtocolDetailsForRun,
